refactor(api): [API-04] :truck: Split integration tests in different files
